Added constraints editing to DatabaseViewer to add/reject loop closures, refine loop closures with ICP, detect more loop closures using the optimized map Added Feature2D::parseParameters()
git-svn-id: http://rtabmap.googlecode.com/svn/trunk/rtabmap@1632 f169173b-cf89-36c8-b27e-44dbe73f0c83

+++ tools/VocabularyComparison/main.cpp
@@ -77,7 +77,7 @@ int main(int argc, char * argv[])
 			{
 				std::getline(file, str);
 				strList = uSplit(str);
-				if(strList.size() == dimension+1)
+				if((int)strList.size() == dimension+1)
 				{
 					//first one is the visual word id
 					std::list<std::string>::iterator iter = strList.begin();
